Rename identity's use_cases/ to commands/
`use_cases/` now holds only commands, so per ADR-0001's end-state naming
it becomes `commands/`, symmetric with `queries/`. The `authentication/`
subpackage and every `*UseCase` class name are unchanged — "use case" is
the layer role; the package encodes the kind.

The package `__init__` drops its re-export list for a docstring, matching
`library` and `tagging`: nothing imported the aggregate, and every caller
already imports the module path directly.

`queries-are-dead-ends` already covers `src.application.*.commands`, so
no contract edit was needed.
