+++ backend/src/application/identity/commands/__init__.py
@@ -1,0 +1,1 @@
+"""Identity command use cases."""
