--- backend/src/application/identity/use_cases/__init__.py
@@ -1,11 +1,0 @@
-from .authentication.authenticate_user_use_case import AuthenticateUserUseCase
-from .authentication.refresh_access_token_use_case import RefreshAccessTokenUseCase
-from .register_user_use_case import RegisterUserUseCase
-from .update_user_use_case import UpdateUserUseCase
-
-__all__ = [
-    "AuthenticateUserUseCase",
-    "RefreshAccessTokenUseCase",
-    "RegisterUserUseCase",
-    "UpdateUserUseCase",
-]
