+++ backend/src/containers/identity.py
@@ -1,15 +1,15 @@
 from dependency_injector import containers, providers
 
-from src.application.identity.queries.get_user_by_id_use_case import GetUserByIdUseCase
-from src.application.identity.use_cases.authentication.authenticate_user_use_case import (
+from src.application.identity.commands.authentication.authenticate_user_use_case import (
     AuthenticateUserUseCase,
 )
-from src.application.identity.use_cases.authentication.logout_use_case import LogoutUseCase
-from src.application.identity.use_cases.authentication.refresh_access_token_use_case import (
+from src.application.identity.commands.authentication.logout_use_case import LogoutUseCase
+from src.application.identity.commands.authentication.refresh_access_token_use_case import (
     RefreshAccessTokenUseCase,
 )
-from src.application.identity.use_cases.register_user_use_case import RegisterUserUseCase
-from src.application.identity.use_cases.update_user_use_case import UpdateUserUseCase
+from src.application.identity.commands.register_user_use_case import RegisterUserUseCase
+from src.application.identity.commands.update_user_use_case import UpdateUserUseCase
+from src.application.identity.queries.get_user_by_id_use_case import GetUserByIdUseCase
 
 
 class IdentityContainer(containers.DeclarativeContainer):
